src/requirements/base.txt
Django==5.2.1
djangorestframework==3.16.0
python-dotenv==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-extensions==4.1
requests==2.32.3
httptools==0.6.4
whitenoise==6.9.0
PyJWT==2.6.0
Pillow==11.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-import-export==4.3.7
django-rest-swagger==2.2.0
drf-yasg==1.21.10
redis==6.1.0
django-redis==5.4.0
boto3==1.38.21
django-storages==1.14.6
shortuuid==1.0.13